tested createSecrets function
